requirements.txt
# ================================
# Alalloy Agent v2.0 后端依赖
# 基于 LangChain 1.x + LangGraph 1.x 官方最佳实践
# ================================

# LLM 框架（核心）
langchain>=1.0.0              # LangChain 核心（create_agent, middleware）
langchain-core>=0.3.0         # LangChain 核心类型
langchain-openai>=0.3.0       # OpenAI 兼容接口（DashScope 百炼 API）
langgraph>=1.0.0              # 工作流编排（StateGraph, stream_mode）
langgraph-checkpoint>=2.0.0   # Checkpointer 基础
langchain-mcp-adapters        # MCP 工具适配器（ONNX + Calphad）

# Web 框架
fastapi>=0.115.0              # Web API 框架
uvicorn[standard]>=0.31.0     # ASGI 服务器（含 WebSocket 支持）
pydantic>=2.9.0               # 数据验证 + structured_output

# HTTP 请求
requests>=2.31.0              # IDME API 调用
httpx>=0.27.0                 # 异步 HTTP（备用）

# JWT 认证
PyJWT[crypto]>=2.8.0          # FerrisKey OIDC Token 验证（RS256 + JWKS）

# 工具
python-dotenv>=1.0.0          # 环境变量管理
loguru>=0.7.0                 # 日志库

# Supabase（会话持久化）
supabase>=2.0.0               # Supabase Python 客户端
langgraph-checkpoint-postgres # AsyncPostgresSaver（Supabase PostgreSQL）
psycopg[binary]>=3.1.0        # PostgreSQL 驱动（asyncpg 底层）
psycopg-pool>=3.1.0           # 连接池
